Moved 2D representation functions to a separate file so it'll be easier to remove them further on. Decided to not fix the flickering
Need to try to get the wall collision checks working
